Aha! The issue is LoaderOptionsPlugin.js
When using the LoaderOptionsPlugin it appears as though there is some
lossy information in Webpack internals.

Specifically, with the same class name and filename like so:

```
styles.css -> .container {}
dir/styles.css -> .container {}
```

We should expect that .container is hashed uniquely and differently from
each other in the output. This is true *until* we add
`LoaderOptionsPlugin` specific for a module.

```
const PostCSS = new Webpack.LoaderOptionsPlugin({
  test: /\.css$/,
  options: {
    postcss: {
      plugins() {
        return [autoprefixer];
      }
    }
  }
});
```

Adding some logging to `loader-utils/index.js` url hashing we can see
that `content` includes the path location without LoaderOptionsPlugin,
but that is lost when using LoaderOptionsPlugin.

```
$ NODE_ENV=production webpack # without LoaderOptionsUtil
> exports.getHashDigest(styles.css+base, undefined, base64, 5);
> exports.getHashDigest(styles.css+header, undefined, base64, 5);
> exports.getHashDigest(styles.css+paragraph, undefined, base64, 5);
> exports.getHashDigest(dir/styles.css+base, undefined, base64, 5);

$ NODE_ENV=development webpack # with LoaderOptionsUtil
> exports.getHashDigest(styles.css+base, undefined, base64, 5);
> exports.getHashDigest(styles.css+header, undefined, base64, 5);
> exports.getHashDigest(styles.css+paragraph, undefined, base64, 5);
> exports.getHashDigest(styles.css+base, undefined, base64, 5);
```

Notice that with LoaderOptionsUtil the final `styles.css+base` *should
be* `dir/styles.css+base`